feat: add modular refactoring tools and config
