+++ src/core/refactoring/analysis_tools.py
@@ -1,0 +1,112 @@
+"""Analysis helpers for the refactoring toolkit."""
+
+from pathlib import Path
+from typing import Dict, Any, List
+
+
+def _get_extraction_recommendations(opportunities: List[str]) -> List[str]:
+    """Get extraction recommendations based on detected opportunities."""
+    recommendations: List[str] = []
+    if "file_size" in opportunities:
+        recommendations.append("Extract large functions into separate modules")
+    if "multiple_classes" in opportunities:
+        recommendations.append("Separate classes into focused modules")
+    if "many_functions" in opportunities:
+        recommendations.append("Group related functions into utility modules")
+    if "mixed_responsibilities" in opportunities:
+        recommendations.append(
+            "Separate imports, classes, and functions into focused modules"
+        )
+    return recommendations
+
+
+def analyze_file_for_extraction(file_path: Path) -> Dict[str, Any]:
+    """Analyze file for module extraction opportunities."""
+    try:
+        with open(file_path, "r", encoding="utf-8", errors="ignore") as f:
+            content = f.read()
+            lines = content.split("\n")
+
+        line_count = len(lines)
+        class_count = content.count("class ")
+        function_count = content.count("def ")
+
+        extraction_opportunities: List[str] = []
+        if line_count > 300:
+            extraction_opportunities.append("file_size")
+        if class_count > 3:
+            extraction_opportunities.append("multiple_classes")
+        if function_count > 10:
+            extraction_opportunities.append("many_functions")
+        if any(
+            keyword in content.lower() for keyword in ["import", "class", "def", "if __name__"]
+        ):
+            if content.count("import") > 5 and class_count > 0 and function_count > 0:
+                extraction_opportunities.append("mixed_responsibilities")
+
+        return {
+            "file_path": str(file_path),
+            "line_count": line_count,
+            "class_count": class_count,
+            "function_count": function_count,
+            "extraction_opportunities": extraction_opportunities,
+            "recommended_actions": _get_extraction_recommendations(extraction_opportunities),
+        }
+    except Exception as e:  # pragma: no cover - safety net
+        return {"error": f"Analysis failed: {e}"}
+
+
+def find_duplicate_files(base_path: Path) -> List[Dict[str, Any]]:
+    """Find duplicate files in the codebase.
+
+    This is a simplified implementation using filename patterns.
+    """
+    duplicates: List[Dict[str, Any]] = []
+    duplicate_patterns = ["api_key_manager.py", "ai_agent_manager.py", "workflow_manager.py"]
+    for pattern in duplicate_patterns:
+        matches = list(base_path.rglob(f"*{pattern}"))
+        if len(matches) > 1:
+            duplicates.append(
+                {
+                    "pattern": pattern,
+                    "files": [str(f) for f in matches],
+                    "duplication_level": "high" if len(matches) > 2 else "medium",
+                }
+            )
+    return duplicates
+
+
+def analyze_architecture_patterns() -> Dict[str, Any]:
+    """Analyze architecture patterns in the codebase.
+
+    This function returns static data in lieu of real analysis.
+    """
+    patterns = [
+        {
+            "name": "BaseManager Inheritance",
+            "description": "Classes inheriting from BaseManager",
+            "count": 15,
+            "quality_score": 85,
+        },
+        {
+            "name": "Module Extraction",
+            "description": "Large files broken into focused modules",
+            "count": 25,
+            "quality_score": 90,
+        },
+        {
+            "name": "Single Responsibility",
+            "description": "Classes following SRP principle",
+            "count": 40,
+            "quality_score": 88,
+        },
+    ]
+    return {
+        "patterns": patterns,
+        "overall_quality_score": sum(p["quality_score"] for p in patterns) / len(patterns),
+        "recommendations": [
+            "Continue BaseManager inheritance pattern",
+            "Extract remaining large modules",
+            "Enforce SRP compliance",
+        ],
+    }
