+++ src/core/refactoring/config.py
@@ -1,0 +1,3 @@
+"""Configuration for the refactoring toolkit."""
+
+DEFAULT_MAX_WORKERS = 4
